build: upgrade webpack-cli, stop building js assets separately (#331)
Version 5 has a breaking change and needs another option to build each entrypoint
separately. This option may be removed in the future and building assets separately
does not allow to deduplicate code between multiple entrypoints, so the build script
is changed to build all the assets together in a simple step.