Reload preview window instead of changing location
This allow to stay at the same place in the previewed page after the render refresh.
For now it is only working with Ctrl + D, not when clicking the button, as the button
is only a simple link. Later the button could be overriden with javascript to allow
the same behaviour